refactor: types 폴더로 타입 이동
