+++ src/apis/place/index.ts
@@ -1,4 +1,3 @@
 export * from './apis';
 export * from './keys';
 export * from './queries';
-export * from './types';
