+++ src/apis/review/index.ts
@@ -1,4 +1,3 @@
 export * from './apis';
-export * from './types';
 export * from './keys';
 export * from './queries';
