+++ src/app/(home)/components/PlaceItem/index.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 
-import type { Place } from '@/apis/place';
+import type { Place } from '@/types/place';
 
 interface PlaceItemProps {
   place: Place;
