+++ src/app/mypage/record/timeline/components/RecordItem/index.tsx
@@ -2,22 +2,22 @@ import { Dot } from '@/app/mypage/components';
 import { format } from 'date-fns';
 import Image from 'next/image';
 
-import type { RecordItem } from '@/types/mypage';
+import type { StudyRecord } from '@/types/record';
 
 interface RecordItemProps {
-  data: RecordItem;
+  data: StudyRecord;
 }
 
 export default function RecordItem({ data }: RecordItemProps) {
-  const { name, content, time } = data;
-  const date = format(new Date(time), "h'시간' m'분'");
+  const { placeName, description, studyDate } = data;
+  const date = format(new Date(studyDate), "h'시간' m'분'");
   return (
     <>
       <div className="flex w-full items-center justify-between gap-4">
         <div className="flex flex-col justify-start gap-0.5">
-          <div className="text-sub1">{name}</div>
+          <div className="text-sub1">{placeName}</div>
           <div className="flex items-center gap-2 text-body2 text-violet/default">
-            <div>{content}</div>
+            <div>{description}</div>
             <Dot />
             <div>{date}</div>
           </div>
