+++ src/app/place/[id]/page.tsx
@@ -22,7 +22,7 @@ import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 
-import type { PlaceConditionType } from '@/app/place/types';
+import type { PlaceConditionType } from '@/types/place';
 
 export default function Page({ params }: { params: { id: string } }) {
   const [isReviewSheetOpen, setIsReviewSheetOpen] = useState(false);
