+++ src/app/place/components/KagongItem/index.tsx
@@ -1,7 +1,7 @@
 import { CONDITION_LIST } from '@/app/place/constants';
 import Image from 'next/image';
 
-import type { PlaceConditionType } from '@/app/place/types';
+import type { PlaceConditionType } from '@/types/place';
 
 interface KagongItemProps {
   type: PlaceConditionType;
