--- src/app/place/types.ts
@@ -1,34 +1,0 @@
-export type LinkType = 'INSTAGRAM' | 'BLOG' | 'WEB';
-
-export type DayType =
-  | 'MONDAY'
-  | 'TUESDAY'
-  | 'WEDNESDAY'
-  | 'THURSDAY'
-  | 'FRIDAY'
-  | 'SATURDAY'
-  | 'SUNDAY';
-
-export type PlaceConditionType = 'CLEAN' | 'WIFI' | 'SEAT' | 'TEMPERATURE' | 'TABLE' | 'QUIET';
-
-export type LineType =
-  | '1'
-  | '2'
-  | '3'
-  | '4'
-  | '5'
-  | '6'
-  | '7'
-  | '8'
-  | '9'
-  | '인천1호선'
-  | '인천2호선'
-  | '신분당'
-  | '경의중앙선'
-  | '경춘선'
-  | '수인분당'
-  | '공항'
-  | '신림선'
-  | '의정부'
-  | '에버라인'
-  | '경강선';
