+++ src/components/Tab/Tab.tsx
@@ -1,4 +1,4 @@
-import { Dispatch, SetStateAction } from 'react';
+import type { Dispatch, SetStateAction } from 'react';
 
 interface TabProps {
   children: React.ReactNode;
