+++ src/types/Image.ts
@@ -1,0 +1,7 @@
+export interface Image {
+  id: number;
+  url: string;
+  width: number;
+  height: number;
+  extension: string;
+}
