+++ src/types/mypage/index.ts
