+++ src/types/search/index.ts
